requirements.txt
python-telegram-bot
supabase
python-dotenv
fpdf2
python-telegram-bot-calendar